flytt hjelpefunksjoner fra hook til egen eøs-utils-fil
